refactor: move pausedWorkflowSchemas to separate file
