+++ backend/pyspur/api/workflow_run.py
@@ -6,7 +6,7 @@ import os
 import re
 from datetime import datetime, timezone
 from pathlib import Path  # Import Path for directory handling
-from typing import Any, Awaitable, Dict, List, Optional, Union, Set, cast
+from typing import Any, Awaitable, Dict, List, Optional, Union, Set
 
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query
 from sqlalchemy.orm import Session
