+++ backend/pyspur/nodes/logic/human_intervention.py
@@ -60,8 +60,9 @@ class HumanInterventionNodeOutput(BaseNodeOutput):
 class HumanInterventionNode(BaseNode):
     """
     A node that pauses workflow execution and waits for human input.
-    The workflow will resume once the required input is provided.
-    All input data is passed through to the output directly after approval.
+
+    When this node is executed, it pauses the workflow until human intervention
+    occurs. All input data is passed through to the output after approval.
     """
     name = "human_intervention_node"
     config_model = HumanInterventionNodeConfig
@@ -76,13 +77,39 @@ class HumanInterventionNode(BaseNode):
 
     async def run(self, input: BaseModel) -> BaseModel:
         """
-        Pass through all input fields directly and pause the workflow.
-        """
-        # Create a flat output structure where fields are directly accessible
-        # This ensures templates like {{HumanInterventionNode_1.input_1}} will work
-        input_dict = input.model_dump()
+        Process input, create an output model, and pause the workflow.
 
-        # Flatten the structure if needed - extract values from any nested node inputs
+        This method passes through input data to the output and raises a
+        PauseException to trigger a workflow pause for human intervention.
+
+        Args:
+            input: The node input data
+
+        Returns:
+            Never returns normally - always raises PauseException
+
+        Raises:
+            PauseException: Always raised to pause workflow execution
+        """
+        # Flatten input structure for easier access in templates
+        flat_input = self._flatten_input(input.model_dump())
+
+        # Create output with all input fields directly accessible
+        output = HumanInterventionNodeOutput(**flat_input)
+
+        # Pause workflow execution
+        raise PauseException(self.name, self.config.message, output)
+
+    def _flatten_input(self, input_dict: dict) -> dict:
+        """
+        Flatten nested input dictionaries to make fields directly accessible.
+
+        Args:
+            input_dict: The input dictionary to flatten
+
+        Returns:
+            A flattened dictionary where nested values are moved to the top level
+        """
         flat_input = {}
         for key, value in input_dict.items():
             if isinstance(value, dict):
@@ -91,8 +118,4 @@ class HumanInterventionNode(BaseNode):
             else:
                 flat_input[key] = value
 
-        # Create output with all input fields directly accessible
-        output = HumanInterventionNodeOutput(**flat_input)
-
-        # Raise PauseException to trigger workflow pause
-        raise PauseException(self.name, self.config.message, output)
+        return flat_input
